[PATCH] 10: 2.5.29-wdt977
Bitops are used with on the timer_alive variable.  Therefore, timer_alive
needs to be "unsigned long" not "int".

+++ drivers/char/wdt977.c
@@ -39,7 +39,7 @@
 
 static	int timeout = DEFAULT_TIMEOUT*60;	/* TO in seconds from user */
 static	int timeoutM = DEFAULT_TIMEOUT;		/* timeout in minutes */
-static	int timer_alive;
+static	unsigned long timer_alive;
 static	int testmode;
 
 MODULE_PARM(timeout, "i");
